refactor(stm32): unify the USB module to ease future changes
